show steps when incomplete
